requirements.txt
django
djangorestframework
dj-database-url
psycopg2-binary
whitenoise[brotli]
gunicorn